refactor(branch_and_bound): make functional.
