+++ pykp/solvers/__init__.py
@@ -3,6 +3,6 @@ PyKP is a package to provide tooling for sampling and solving instances of the 0
 """
 
 from .solver import Solver
-from .branch_and_bound import BranchAndBound
+from ._branch_and_bound import branch_and_bound
 from .greedy import Greedy
 from .mzn_gecode import MznGecode
